Reorganize the branch with the file structure in main.
Might have done wrong with the branch, but let's keep it.
